Cholesky Module
Aggiunto .gitignore per ignorare file compilati di python.
Creato modulo per la decomposizione di cholesky.
